GH-25 rename rosbag2_test_commons -> rosbag2_test_common
