+++ rosbag2_test_common/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(rosbag2_test_commons)
+project(rosbag2_test_common)
 
 # Default to C99
 if(NOT CMAKE_C_STANDARD)
