+++ rosbag2_test_common/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="2">
-  <name>rosbag2_test_commons</name>
+  <name>rosbag2_test_common</name>
   <version>0.0.0</version>
   <description>Commonly used test helper classes and fixtures for rosbag2</description>
   <maintainer email="karsten.knese@googlemail.com">Karsten Knese</maintainer>
