+++ rosbag2_tests/package.xml
@@ -15,7 +15,7 @@
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
   <test_depend>rosbag2_storage_default_plugins</test_depend>
-  <test_depend>rosbag2_test_commons</test_depend>
+  <test_depend>rosbag2_test_common</test_depend>
   <test_depend>rosbag2_storage</test_depend>
   <test_depend>ros2bag</test_depend>
   <test_depend>rclcpp</test_depend>
