+++ rosbag2_tests/test/rosbag2_tests/record_fixture.hpp
@@ -35,13 +35,13 @@
 #include "test_msgs/msg/static_array_primitives.hpp"
 #include "test_msgs/message_fixtures.hpp"
 #include "rosbag2_storage_default_plugins/sqlite/sqlite_storage.hpp"
-#include "rosbag2_test_commons/temporary_directory_fixture.hpp"
-#include "rosbag2_test_commons/publisher_manager.hpp"
-#include "rosbag2_test_commons/memory_management.hpp"
+#include "rosbag2_test_common/temporary_directory_fixture.hpp"
+#include "rosbag2_test_common/publisher_manager.hpp"
+#include "rosbag2_test_common/memory_management.hpp"
 
 using namespace ::testing;  // NOLINT
 using namespace std::chrono_literals;  // NOLINT
-using namespace rosbag2_test_commons;  // NOLINT
+using namespace rosbag2_test_common;  // NOLINT
 
 #ifdef _WIN32
 using ProcessHandle = HANDLE;
