+++ rosbag2_tests/test/rosbag2_tests/test_rosbag2_play_end_to_end.cpp
@@ -32,11 +32,11 @@
 #include "test_msgs/msg/static_array_primitives.hpp"
 #include "test_msgs/message_fixtures.hpp"
 #include "rosbag2_storage_default_plugins/sqlite/sqlite_storage.hpp"
-#include "rosbag2_test_commons/subscription_manager.hpp"
+#include "rosbag2_test_common/subscription_manager.hpp"
 
 using namespace ::testing;  // NOLINT
 using namespace std::chrono_literals;  // NOLINT
-using namespace rosbag2_test_commons;  // NOLINT
+using namespace rosbag2_test_common;  // NOLINT
 
 class EndToEndTestFixture : public Test
 {
