+++ rosbag2_transport/package.xml
@@ -17,7 +17,7 @@
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
   <test_depend>test_msgs</test_depend>
-  <test_depend>rosbag2_test_commons</test_depend>
+  <test_depend>rosbag2_test_common</test_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
